requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
scikit-learn
mysql-connector-python